修复代码截图控制器中的返回类型，将截图数量的返回类型从Integer更改为Long；更新评审分配控制器中的状态设置，直接使用状态枚举；在通知实体中添加了构造函数和构建器；新增截图重新排序功能，并优化了相关服务中的截图数量获取逻辑；修正通知服务中今日和本周通知数量的返回类型为Long；更新通知渠道服务中的异常处理，确保用户信息获取的安全性。

+++ code-review-user/src/main/java/com/company/codereview/user/controller/CodeScreenshotController.java
@@ -249,9 +249,9 @@ public class CodeScreenshotController {
     @GetMapping("/review-record/{reviewRecordId}/count")
     @ApiOperation("获取评审记录的截图数量")
     @PreAuthorize("hasRole('DEVELOPER') or hasRole('TEAM_LEADER') or hasRole('ARCHITECT')")
-    public ResponseResult<Integer> getScreenshotCount(@PathVariable Long reviewRecordId) {
+    public ResponseResult<Long> getScreenshotCount(@PathVariable Long reviewRecordId) {
         
-        int count = codeScreenshotService.getScreenshotCount(reviewRecordId);
+        Long count = codeScreenshotService.getScreenshotCount(reviewRecordId);
         
         return ResponseResult.success(count);
     }
